Made seperate folders in projects assets, css & js to house my elements due to I have multiple stylesheets that work together, a music player and assets ranging from audio to pictures.
Still working on making it more responsive to different screen sizes. Currently its optimized for big screens such as monitors, lcd and laptops. Might add an extra element of an introduction video the product.

Also during the week will be working on the form abit to get it to log my php messages to a database or atleast the alerts to say messages have been sent

Also just remembered have to work on my music player to show the icons when on pause and playing next to the equalizer near the menu button. Same time hopefully will make the equalizer react to the music and give an animation. Gonna look at preloader spinners to add as well to give it that interest when its loading